% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify_string.R
\name{prettify_names}
\alias{prettify_names}
\title{Wrap of prettify_string for names}
\usage{
prettify_names(x, ...)
}
\arguments{
\item{x}{A dataframe, a tibble, a named vector, a named list.}

\item{...}{Custom replacement. See prettify_strings.}
}
\value{
Returns x with clean names.
}
\description{
Remove:
\itemize{
\item punctuation,
\item special characters,
\item spaces and tabs,
\item accents
}

Force everything to lower.
}
\examples{

head(iris)
head(prettify_names(iris))

}
