% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify_string.R
\name{prettify_string}
\alias{prettify_string}
\title{Prettify strings}
\usage{
prettify_string(x, ...)
}
\arguments{
\item{x}{A string or a vector of strings.}

\item{...}{Custom replacement. See examples.}
}
\value{
Returns a clean x
}
\description{
Remove:
\itemize{
\item punctuation,
\item special characters,
\item spaces and tabs,
\item accents
}

Force everything to lower
}
\examples{

# remove all symbols and accents
x <- "a1~!#$€\%^&*|__+:____()[]{}<>?,./;:'-=òìàèéù dASFGHRV"
y <- c("a1~!#$€\%^&*|__+:____()[]{}<>?,./;:'-=òìàèéù dASFGHRV",
       "CIAAAOOOO")

prettify_string(x)
prettify_string(y)

# custom replacement
prettify_string(x = "5\%_in_€")
prettify_string(x = "5\%_in_€", "\%" = "perc")
prettify_string(x = "5\%_in_€", "\%" = "perc", "€" = "eur")


}
