% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify_string.R
\name{prettify_title}
\alias{prettify_title}
\title{Wrap of prettify_string for titles}
\usage{
prettify_title(x, ..., replace = " ")
}
\arguments{
\item{x}{A string or a vector of strings.}

\item{...}{Custom replacement. See prettify_strings.}

\item{replace}{String to use instad of the underscore. Default blank.}
}
\value{
Returns a clean x as printable title.
}
\description{
Remove:
\itemize{
\item punctuation,
\item special characters,
\item spaces and tabs,
\item accents
}

Force everything to title with spaces between words.
}
\examples{

x <- "a1~!#$\%^&*|__+:____()[]{}<>?,./;:'-=òìàèéù dASFGHRV"
y <- c("a1~!#$\%^&*|__+:____()[]{}<>?,./;:'-=òìàèéù dASFGHRV",
       "CIAAAOOOO")

prettify_title(x)
prettify_title(y)

}
